Adding naive bayes service unit tests

+++ naivebayes/src/main/java/com/haines/ml/rce/naivebayes/model/Probability.java
@@ -9,7 +9,7 @@ public class Probability {
 	public Probability(int outcomes, int totals){
 		this.outcomes = outcomes;
 		this.totals = totals;
-		this.probability = outcomes / totals;
+		this.probability = outcomes / (double)totals;
 	}
 	
 	public double getProbability(){
